requirements.txt
vk_api
random
time
json
pymysql.cursors
re
